On index.php, added the include for phpcache, as well as added in a check for an XMLREQUEST so that the loading of menus can later be skipped for AJAX calls.
System include phpcache is a newly added mechanism for caching particular component/action/query combinations to enhance Pines' performance on static pages. Caching is managed by the component com_cache and is easy to implement. A default sample_cacheoptions.php file is included and a copy of it is made through com_cache when the manager uses the default directives. 

get_component_classes and get_component_inits are files to be executed from the terminal, possibly in a bash script when updating repos - and it basically consolidates the chore of reading every class file of every component and every init file into only 1 file, respectively. If the two generated files do not exist, Pines will behave in the traditional way and not use them. The result when these files are implemented is reduced CPU intensive work per every call through index.php.